Move InterpolationMeasurmentProviderTest from simulator to app module
