+++ org.envirocar.app/src/org/envirocar/app/handler/InterpolationMeasurementProvider.java
@@ -203,8 +203,8 @@ public class InterpolationMeasurementProvider extends AbstractMeasurementProvide
      * @param endTimestamp    the timestamp of the lend
      * @return the interpolated value
      */
-    protected Double interpolateTwo(Number start, Number end, long targetTimestamp,
-                                    long startTimestamp, long endTimestamp) {
+    public Double interpolateTwo(Number start, Number end, long targetTimestamp,
+                                 long startTimestamp, long endTimestamp) {
         if (start == null && end == null) {
             return null;
         }
